welcome/workshop/part-3-product-consumer/requirements.txt
mysql-connector
pika